T3: Calls to database + database
implemented an early version of the database. check the code for it at DB code.txt
IF YOU MAKE ANY CHANGES TO THE DB MAKE THE CHANGES TO THE DB CODE TOO!!! pls or leave it to me as if it were nothing.

Also implemented the first part of the T3 server, you can access the database, and read(so far) data thats stored